% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorityMesh.R
\name{autoplot.fm_mesh_2d_intensity}
\alias{autoplot.fm_mesh_2d_intensity}
\title{Plot a Mesh Object}
\usage{
\method{autoplot}{fm_mesh_2d_intensity}(object, ...)
}
\arguments{
\item{object}{A \code{fm_mesh_2d_intensity} object created from the
\code{\link{priorityMesh}} function.}

\item{...}{A series of named parameters prefixed with \code{"intensity."},
\code{"mesh."}, \code{"initialLocs."}, \code{"boundary."} to pass parameters
to the \code{\link[ggplot2]{geom_sf}} or
\code{\link[tidyterra]{geom_spatraster}} geometry functions that determine
the plotting behaviour of the intensity surface, mesh, initial triangulation
nodes, and boundary polygon respectively. Is is also possible to specify
other arguments that control the look of elements related to the aesthetic
properties of the plot, see 'Details' below.}
}
\value{
A \code{ggplot} object
}
\description{
Function to plot a \code{fm_mesh_2d_intensity} object.  This
object type is what is returned from the \code{\link{priorityMesh}} function.
}
\details{
In addition to the arguments outlined above, the user may also set the
following optional arguments that control the look of elements related to the
aesthetic properties of the plot:
\describe{
\item{\code{low.intensity.col}}{A character scalar containing the colour to
use for the cells with the lowest intensity values in the intensity surface.
}
\item{\code{high.intensity.col}}{A character scalar containing the colour to
use for the cells with the highest intensity values in the intensity
surface.}
\item{\code{na.intensity.col}}{A character scalar containing the colour to
use for the cells with an NA value in the intensity surface.}
\item{\code{provided.initialLocs.col}}{A character scalar containing the
colour to use to plot the points that were provided by the user as initial
triangulation nodes.}
\item{\code{generated.initialLocs.col}}{A character scalar containing the
colour to use to plot the points that were generated by the
\code{\link{priorityMesh}} function.}
}
}
\seealso{
\code{\link[fmesher]{fm_mesh_2d_inla}} \code{\link{priorityMesh}}
\code{\link[ggplot2]{geom_sf}} \code{\link[tidyterra]{geom_spatraster}}
}
\author{
Joseph D. Chipperfield, \email{joseph.chipperfield@nmbu.no}
}
