% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcInternals.R
\name{centreScaleCovariates}
\alias{centreScaleCovariates}
\title{Centre and Scale Covariates}
\usage{
centreScaleCovariates(inData, centreCovs = TRUE, scaleCovs = TRUE)
}
\arguments{
\item{inData}{A data.frame or object coercible to a data.frame containing the
variables to centre and scale.}

\item{centreCovs}{A logical scalar denoting whether the fixed effects in the
model should be centred before the analysis: each covariate element is
subtracted by its mean. \code{centreCovs} can also be a function with one
argument that is a vector of covariate values. In this case the variable is
instead centred around the output of this function.}

\item{scaleCovs}{A logical scalar denoting whether the fixed effects in the
model should be scaled before the analysis: each covariate element is divided
by its standard deviation. \code{scaleCovs} can also be a function with one
argument that is a vector of covariate values. In this case the variable is
instead scaled around the output of this function.}
}
\value{
A data.frame object containing the scaled and centred values of the
covariates. In addition the object is given two attributes:
\describe{
\item{centreFactor}{A named numeric vector of length equal to the number of
columns in \code{inData} that contains the number used to centre the
variable or \code{NA} if that variable remains uncentred}
\item{scaleFactor}{A named numeric vector of length equal to the number of
columns in \code{inData} that contains the number used to scale the
variable or \code{NA} if that variable remains unscaled}
}
}
\description{
Centre and Scale Covariates
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
