% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownload.R
\name{check_gbifEmail}
\alias{check_gbifEmail}
\title{Find GBIF Email}
\usage{
check_gbifEmail(val)
}
\arguments{
\item{val}{A character scalar containing the GBIF username}
}
\value{
A user name for GBIF
}
\description{
A function that looks in the common places for a GBIF email
address if the address is not passed as an argument to the function:
\code{'GBIF_EMAIL'} system variable or from a call to
\code{getOption('gbif_email')}.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
