% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownload.R
\name{check_gbifPwd}
\alias{check_gbifPwd}
\title{Find GBIF Password}
\usage{
check_gbifPwd(val)
}
\arguments{
\item{val}{A character scalar containing the GBIF password}
}
\value{
A password for GBIF
}
\description{
A function that looks in the common places for a GBIF password
if the password is not passed as an argument to the function:
\code{'GBIF_PWD'} system variable or from a call to
\code{getOption('gbif_pwd')}.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
