% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownload.R
\name{check_gbifUser}
\alias{check_gbifUser}
\title{Find GBIF Username}
\usage{
check_gbifUser(val)
}
\arguments{
\item{val}{A character scalar containing the GBIF username}
}
\value{
A user name for GBIF
}
\description{
A function that looks in the common places for a GBIF username
if the username is not passed as an argument to the function:
\code{'GBIF_USER'} system variable or from a call to
\code{getOption('gbif_user')}.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
