% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcInternals.R
\name{customError}
\alias{customError}
\title{Function to Define a Custom Error Distribution}
\usage{
customError(
  nimbleLikeli,
  simulate,
  nimblePrior = list(),
  nimbleConstants = list(),
  link = "identity",
  discrete = FALSE,
  elementWise = FALSE
)
}
\arguments{
\item{nimbleLikeli}{A function that returns a string containing the text for
a specification of the likelihood in NIMBLE. The function must take three
arguments: the first argument is a string that contains the name of the node
that contains the expectation of the linear model, the second argument is a
string that contains the node containing the data, and the third argument is
an optional suffix that can be added to all the names of the nodes in the
linear model.}

\item{simulate}{A function that takes the expected value as its first
argument along with named arguments equal to those defined in
\code{nimblePrior} and \code{nimbleConstants} and then draws a number of
random draws from the error distribution with the specified parameterization.}

\item{nimblePrior}{A list with named elements corresponding to each
parameter used in the error distribution and with each element being a
function that returns a string containing a prior specification for that
parameter in NIMBLE code. Each function element takes ones argument: an
optional suffix that can be added to all the names of the nodes in the linear
model.}

\item{nimbleConstants}{A list with named elements corresponding to each
constant value used in the error distribution and with each element being a
function that returns a numeric vector containing the values for the
constants. Each function is passed the complete set of parameters that are
passed to the \code{\link{modelDefinitionToNIMBLE}} function. Alternatively,
the constants can just be values as they would be used by the
\code{\link[nimble]{nimbleModel}} function.}

\item{link}{Either a character vector containing the names of supported
link functions (see \code{unique(unlist(errorFamilies()))} for a list of
natively-supported link functions) or a list of named elements and each
element is the output of the \code{\link{customLink}} function. The first
element of \code{link} will be used as the default link function for the
distribution.}

\item{discrete}{A logical scalar denoting whether the distribution is defined
for a discrete variable or not.}

\item{elementWise}{A logical scalar denoting whether the \code{nimbleLikeli}
function defines the likelihood for one data point at a time (using a loop
in the model specification) or defines the entire data vector as a
multivariate output from a joint probability distribution.}
}
\value{
A list containing the following named elements:
\describe{
\item{\code{link}}{A list of link functions that are valid for this error
distribution. Each element is a list containing the output from a call to
the \code{\link{customLink}} function}
\item{\code{nimbleLikeli}}{A version of the \code{nimbleLikeli} function
encapsulated in such a way as to be usable by the package}
\item{\code{nimblePrior}}{A version of the \code{nimblePrior} list with each
element encapsulated in such as a way as to by usable by the package}
\item{\code{nimbleConstants}}{A version of the \code{nimbleConstants} list
with each element encapsulated in such a way as to be usable by the package}
\item{\code{elementWise}}{A copy of the input argument \code{elementWise}}
\item{\code{discrete}}{A copy of the input argument \code{discrete}}
\item{\code{simulate}}{A version of the \code{simulate} function
encapsulated in such a way as to be usable by the package}
}
}
\description{
A function that allows for the definition of user-specified
error distributions for use in generalized linear models
}
\seealso{
\code{\link[nimble]{nimbleCode}} \code{\link{customLink}}
\code{\link{errorFamilies}} \code{\link{modelDefinitionToNIMBLE}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
