% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownload.R
\name{datasetplot}
\alias{datasetplot}
\title{Plot GBIF Occurrence Data}
\usage{
datasetplot(object, ..., titleLabelWidth = 40)
}
\arguments{
\item{object}{An object that inherits from class \code{gbifOcc} (such as that
produced by the \code{\link{gbifOccDownload}} function).}

\item{...}{Named arguments passed to constituent plotting functions.
Arguments that are of the form \code{points.parameter} will be passed to the
\code{\link[ggplot2]{geom_sf}} function that determines the plotting
behaviour of the occurrence records (without the \code{points.} prefix). If
arguments of the form \code{boundary.parameter} are provided then a boundary
layer is added to the plot (with data drawn from the \code{boundary.data}
argument) and these arguments are similarly passed to the corresponding
\code{\link[ggplot2]{geom_sf}} function (without the \code{boundary.}
prefix). Argument with the prefixes \code{yearbar.} and \code{datasetbar.}
are passed to the \code{\link[ggplot2]{geom_bar}} functions that plot the
dataset occurrences counted by year and source dataset respectively.
Arguments with the \code{colour.} prefix are passed to the
\code{\link[ggplot2]{scale_colour_discrete}} and
\code{\link[ggplot2]{scale_fill_discrete}} functions that define the colour
palettes in the constituent plots. Finally, arguments with \code{arrange.}
prefix are passed to the \code{\link[gridExtra]{arrangeGrob}} function that
arranges the constituent plots.}

\item{titleLabelWidth}{An integer scalar giving the number of characters to
aim for when wrapping text associated with the titles of dataset names in
legends and axis labels.}
}
\value{
A \code{datasetplotGrob} object.
}
\description{
Function to visualise the data associated with a GBIF occurrence
download.
}
\seealso{
\code{\link{gbifOccDownload}}, \code{\link[ggplot2]{autoplot}},
\code{\link[ggplot2]{geom_sf}}, \code{\link[gridExtra]{arrangeGrob}},
\code{\link[ggplot2]{scale_colour_discrete}},
\code{\link[ggplot2]{scale_fill_discrete}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
