% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleExtensions.R
\name{dbetabin}
\alias{dbetabin}
\title{Probability Density of the Beta-Binomial Distribution ----}
\usage{
dbetabin(x, shape1, shape2, size, log = 0)
}
\arguments{
\item{x}{A scalar integer containing the value to calculate a density for}

\item{shape1}{A scalar value containing the first shape parameter of the composite
beta distribution}

\item{shape2}{A scalar value containing the second shape parameter of the composite
beta distribution}

\item{size}{A scalar integer containing the number of trials from the composite
binomial distribution}

\item{log}{If \code{TRUE} then return the log density}
}
\value{
A scalar containing the probability density
}
\description{
Calcualte the probability density of the beta-binomial distribution.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
