% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_plateau.R
\name{dplateau}
\alias{dplateau}
\alias{dplateaumags}
\alias{rplateaumags}
\title{Distribution for the magnitude of ascending and descending plateau magnitude coefficients}
\usage{
dplateaumags(x, prec, rho, log = 0)

rplateaumags(n, prec, rho)
}
\arguments{
\item{x}{Numeric vector of length 2 with the first element containing the value of the
ascending magnitude coefficient and the second element containing the value of the
descending magnitude coefficient}

\item{prec}{Scalar double containing the precision hyperparameter to control the scale of
the magnitude coefficients}

\item{rho}{Scalar double containing the correlation coefficient of the ascending and
descending magnitude coefficients}

\item{log}{Logical scalar which if \code{TRUE} (default) results in the return of the log
probability density.  Otherwise output will be on the natural scale}

\item{n}{Number of random draws.  Currently only \code{n = 1} is supported, but the argument
exists for standardization of "\code{r}" functions.}
}
\value{
For \code{dplateaumags} the probability density (on the log or natural scale
depending on the value used for \code{log}) is return as a numeric scalar.  For
\code{rplateaumags} a vector of length two is returned with the first element containing
the magnitude coefficient for the ascending portion and the second element containing the
magnitude coefficient for the descending portion.
}
\description{
\code{dplateaumags} and \code{rplateaumags} provide the required information for
defining the distribution for the magnitude of ascending and descending plateau magnitude
coefficients.  \code{dplateaumags} defines the probability density function and
\code{rplateaumags} defines a function for generating random samples from this distribution.
They can be used directly from R or in \code{nimble} models.
}
\seealso{
\code{\link{h.plateau}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
