% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcInternals.R
\name{errorFamilies}
\alias{errorFamilies}
\title{Error distributions supported in PaGAn}
\usage{
errorFamilies()
}
\value{
A list object containing one named entry for each natively supported
error distributions. Each named entry of the list is a character vector
containing the valid link functions to use with the given error family
}
\description{
\code{errorFamilies} is function that returns the families of error
distributions supported natively in PaGAn's model processing routines
}
\examples{
# Retrieve the names of the error distribution supported in PaGAn
names(errorFamilies())
# Retrieve the link functions that can be used with the "gaussian" error
# family
errorFamilies()[["gaussian"]]
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
