% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialEcosystemState.R
\name{fitMultinomialEcosystemState}
\alias{fitMultinomialEcosystemState}
\title{Fit the Multinomial Ecosystem State Model}
\usage{
fitMultinomialEcosystemState(
  stateValModels,
  stateProbModels,
  statePrecModels,
  inputData,
  numStates = NULL,
  stateValError = gaussian,
  mcmcIters = 10000,
  mcmcBurnin = 5000,
  mcmcChains = 4,
  mcmcThin = 1
)
}
\arguments{
\item{stateValModels}{A formula describing the regression relationship between the mean
ecosystem state value and the covariates for all ecosystem states, or a list of formulas
with each element giving the regression relationship between the mean ecosystem state
value and the covariates for each ecosystem state (ordered according to intercept of the
ecosystem state value on the y-axis).  The ecosystem state variable must be given on the
left-hand side of the formula.}

\item{stateProbModels}{A formula describing the regression relationship between the probability
of the existence of ecosystem state and the covariates for all ecosystem states, or a list
of formulas with each element giving the regression relationship between the probability
of the existence of ecosystem state and the covariates for each ecosystem state (ordered
according to intercept of the ecosystem state value on the y-axis).  The ecosystem state
variable must be given on the left-hand side of the formula.}

\item{statePrecModels}{A formula describing the regression relationship between the variance
in the ecosystem state value and the covariates for all ecosystem states, or a list of
formulas with each element giving the regression relationship between the variance in the
ecosystem state value and the covariates for each ecosystem state (ordered according to
intercept of the ecosystem state value on the y-axis).  The ecosystem state variable must
be given on the left-hand side of the formula.}

\item{inputData}{A data frame containing the covariate information and ecosystem state
variable.}

\item{numStates}{An integer scalar containing the number of stable states in the ecosystem.
If any of the \code{stateValModels}, \code{stateProbModels}, or \code{statePrecModels} parameters
is a list then \code{numStates} can be omitted and is therefore set to the length of the
list.}

\item{stateValError}{A description of the error distribution and link function to be used
in the model describing the ecosystem state value.  This can be from the \link[stats]{family}
specification or \code{character} scalar with the following possible values: \code{"gaussian"},
\code{"gamma"}, \code{"beta"}, \code{"negbinomial"}, or \code{"betabinomial"}.}

\item{mcmcChains}{An integer scalar giving the number of MCMC chains to use.}

\item{mcmcThin}{An integer scalar giving the thinning frequency in the MCMC chains.  For example,
a value of \code{4} results in every fourth values being retained.}

\item{mcmcIter}{An integer scalar providing the number of post-burn-in samples to draw from the
MCMC sampler (per chain).}

\item{mcmcBurnIn}{An integer scalar providing the number of MCMC samples to use for the
adaption or burn-in portion of the process (per chain).}
}
\value{
A list containing the following components:
\itemize{
\item{\code{mcmcSamples}}{An \link[coda]{mcmc} object if \code{mcmcChains == 1} or \link[coda]{mcmc.list}
object if \code{mcmcChains > 1} containing the sampled values}
\item{\code{compiledModel}}{An R interface object containing an interface for the compiled NIMBLE model.
This is the same output as the \link[nimble]{compileNimble} function applied to the model object}
\item{\code{modelText}}{A character scalar containing the text of the model specification in
the NIMBLE BUGS dialect}
\item{\code{modelCode}}{A \code{nimbleCode} object containing the model specification}
\item{\code{constants}}{A list containing the constants provided to NIMBLE}
\item{\code{data}}{A list containing the data provided to NIMBLE}
\item{\code{errorModel}}{A factor containing the error model used for the specification of
the error distribution for the ecoystem state variable}
\item{\code{linkFunction}}{A factor containing the link function used in the specification
of the ecosystem state variable models}
\item{\code{initialValues}}{A list containing potential initial values used for each of the stochastic
nodes in the NIMBLE model specification}
}
}
\description{
This function takes a set of model specifications for the three sub-model
components of the multinomial ecosystem state model and fits them to a data set.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
