% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownload.R
\name{gbifDefaultCRS}
\alias{gbifDefaultCRS}
\title{Retrieve GBIF's Default CRS}
\usage{
gbifDefaultCRS()
}
\value{
A \code{crs} object containing GBIF's coordinate reference system
information
}
\description{
A function that returns the default coordinate reference system
information that is used by GBIF's API
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
