% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownload.R
\name{gbifOccDownload}
\alias{gbifOccDownload}
\title{Download Occurrence Records from GBIF}
\usage{
gbifOccDownload(
  user,
  pwd,
  email,
  curlopts = list(),
  body = NULL,
  pingtime = 5,
  tmploc = tempdir(),
  assf = TRUE,
  ...
)
}
\arguments{
\item{user}{A character string containing a valid username associated with a
GBIF account. If \code{NULL} then \code{Sys.getenv('GBIF_USER')} and
\code{getOptions('gbif_user')} are searched for a value instead.}

\item{pwd}{A character string containing a valid password associated with a
GBIF account. If \code{NULL} then \code{Sys.getenv('GBIF_PWD')} and
\code{getOptions('gbif_pwd')} are searched for a value instead.}

\item{email}{A character string containing a valid email associated with a
GBIF account. If \code{NULL} then \code{Sys.getenv('GBIF_EMAIL')} and
\code{getOptions('gbif_email')} are searched for a value instead.}

\item{curlopts}{List of named curl options passed on to
\code{\link[rgbif]{occ_download}}.}

\item{body}{A character scalar containing a syntactically valid JSON query
to pass to the GBIF occurrence download API. If this is \code{NULL} then the
arguments in \code{...} are instead passed to the
\code{\link{jsonOccFormulation}} to generate a query.}

\item{pingtime}{An integer scalar giving the number of seconds to wait
between pings to GBIF's servers to query the status of the data download.}

\item{tmploc}{A character scalar providing the location to store temporary
files created in the download (that are then deleted on completion).}

\item{assf}{A logical scalar that if \code{TRUE}, converts the output of the
occurrence record download to an \code{sf} object.}

\item{...}{If \code{body} is \code{NULL} then these parameters are passed to
the \code{\link{jsonOccFormulation}} to generate a query.}
}
\value{
An object of type \code{gbifOcc} that inherits from an \code{sf}
object if \code{assf} is \code{TRUE} but otherwise directly inherits from
a \code{data.frame}. In addition the object will have the following
attributes:
\describe{
\item{\code{DOI}}{A character scalar giving the DOI allocated to the data
download.}
\item{\code{citation}}{A character scalar giving the citation to use when
referring to the dataset in published literature.}
\item{\code{query}}{A \code{jsonlite} object containing the full query used
to generate the download.}
\item{\code{dataset}}{A data frame containing information on the constituent
datasets used in the occurrence download.}
}
}
\description{
Function to query the
\href{https://techdocs.gbif.org/en/data-use/api-downloads.html}{occurrence download API}
of \href{https://www.gbif.org/}{GBIF} and retrieve matching records.
}
\seealso{
\code{\link[rgbif]{name_lookup}}, \code{\link{taxonLookup}},
\code{\link{jsonOccFormulation}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
