% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_iid.R
\name{h.iid}
\alias{h.iid}
\title{Define an IID Random Hierarchical Component in a Linear Model}
\usage{
h.iid(
  var,
  effName = NULL,
  centreCovs = FALSE,
  scaleCovs = FALSE,
  suffix = "",
  iidPrecPrior = "dgamma(0.001, 0.001)"
)
}
\arguments{
\item{var}{The variable around which the hierarchical effect will be defined.
This can be a \code{data.frame}, \code{matrix}, or a vector containing the
different levels to deine the effect over.}

\item{effName}{A character scalar giving a name for the hierarchical effect
being defined and used as name for the appropriate nodes.}

\item{centreCovs}{A logical scalar denoting whether the fixed effects in the
model should be centred before the analysis: each covariate element is
subtracted by its mean. \code{centreCovs} can also be a function with one
argument that is a vector of covariate values. In this case the variable is
instead centred around the output of this function.}

\item{scaleCovs}{A logical scalar denoting whether the fixed effects in the
model should be scaled before the analysis: each covariate element is divided
by its standard deviation. \code{scaleCovs} can also be a function with one
argument that is a vector of covariate values. In this case the variable is
instead scaled around the output of this function.}

\item{suffix}{A character scalar giving an additional suffix applied to
all elements created in the hierarchical model specification.}

\item{iidPrecPrior}{A character scalar containing the NIMBLE code that
determines of the distribution of prior specification of the precision
parameter in the iid random effect.}
}
\value{
A list element with the following named elements:
\describe{
\item{\code{name}}{A character scalar containing the name of the
hierarchical effect and is used as a name for intermediary variables}
\item{\code{code}}{A character scalar containing the NIMBLE code specifying
the hierarchical effect (and will be passed to
\code{\link[nimble]{nimbleCode}})}
\item{\code{constants}}{A list containing named elements corresponding to
the variables used as constants needed for the hierarchical effect in
\code{\link[nimble]{nimbleModel}}}
\item{\code{data}}{A list containing named elements corresponding to the
data nodes used for the hierarchical effect in
\code{\link[nimble]{nimbleModel}}}
\item{\code{inits}}{A named list of starting values for model variables used
in the hierarchical effect and passed to \code{\link[nimble]{nimbleModel}}}
\item{\code{monitors}}{The nodes of the hierarchical effect to
monitor in the MCMC and passed to \code{\link[nimble]{configureMCMC}}}
\item{\code{monitors2}}{The nodes of the hierarchical effect to
monitor in the supplemental chain monitor in the MCMC and passed to
\code{\link[nimble]{configureMCMC}}}
\item{\code{initCode}}{A list of language objects to run upon initialisation
of the NIMBLE instance (see \code{\link{mcmcNIMBLERun}})}
\item{\code{exitCode}}{A list of language objects to run upon completion of
the NIMBLE instance (see \code{\link{mcmcNIMBLERun}})}
\item{\code{runTimeGlobal}}{A list of objects to pass to be compied into
each environment of each NIMBLE instance (see \code{\link{mcmcNIMBLERun}})}
\item{\code{projFunc}}{A function as produced by the
\code{\link[nimble]{nimbleFunction}} function that maps the random variables
defined by the hierarchical model to the data. If \code{NULL} then it is
assumed the effect is defined with the same structure as the data already}
}
}
\description{
Function to specify an iid random hierarchical component within
a Bayesian model specification. This function is not usually called directly
but instead called during processing of the \code{\link{h}} terms in a
hierarchical linear model specification.
}
\seealso{
\code{\link[nimble]{nimbleCode}}, \code{\link{mcmcNIMBLERun}},
\code{\link[nimble]{configureMCMC}}, \code{\link[nimble]{nimbleFunction}},
\code{\link{h}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
