% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_lasso.R
\name{h.lasso}
\alias{h.lasso}
\title{Specify a LASSO Hierarchical Effect}
\usage{
h.lasso(
  var,
  centreCovs = TRUE,
  scaleCovs = TRUE,
  suffix = "",
  lassoRatePrior = "dgamma(0.001, 0.001)"
)
}
\arguments{
\item{var}{The variable around which the hierarchical effect will be defined}

\item{centreCovs}{A logical scalar denoting whether the fixed effects in the
model should be centred before the analysis: each covariate element is
subtracted by its mean. \code{centreCovs} can also be a function with one
argument that is a vector of covariate values. In this case the variable is
instead centred around the output of this function.}

\item{scaleCovs}{A logical scalar denoting whether the fixed effects in the
model should be scaled before the analysis: each covariate element is divided
by its standard deviation. \code{scaleCovs} can also be a function with one
argument that is a vector of covariate values. In this case the variable is
instead scaled around the output of this function.}

\item{suffix}{A character scalar that will be appended to all variables used
in the NIMBLE code (including constants and data)}

\item{lassoRatePrior}{A character scalar containing the NIMBLE code (that
will be processed in \code{\link[nimble]{nimbleCode}}) that defines the
prior for the rate parameter of the Laplace distribution that determines
the shrinkage}
}
\description{
Function used to specify a LASSO hierachical effect
}
\seealso{
\code{\link[nimble]{nimbleCode}}, \code{\link{h}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
