% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{h}
\alias{h}
\title{Define a Hierarchical Component in a Linear Model}
\usage{
h(...)
}
\arguments{
\item{...}{Named arguments to be passed to the hierarchical model
specification}
}
\description{
Function to specify a hierarchical component with a Bayesian
model specification
}
\seealso{
\code{\link[nimble]{nimbleCode}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
