% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcInternals.R
\name{mergeNIMBLEInputs}
\alias{mergeNIMBLEInputs}
\title{Merge Automatically-Generated NIMBLE Arguments with User Inputs}
\usage{
mergeNIMBLEInputs(autoArgs, ...)
}
\arguments{
\item{autoArgs}{A list of model settings as produced by the output of the
\code{\link{modelDefinitionToNIMBLE}} function}

\item{...}{User-supplied arguments to NIMBLE component functions}
}
\description{
An internal function to take input entered by the user using
the NIMBLE interface functions defined in PaGAn (such as
\code{glmmble}) and merge it with the NIMBLE arguments that are
automatically generated by the package
}
\seealso{
\code{\link{modelDefinitionToNIMBLE}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
