% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcInternals.R
\name{nimbleParameters}
\alias{nimbleParameters}
\title{Processing of NIMBLE parameters}
\usage{
nimbleParameters(..., warnNotFound = FALSE)
}
\arguments{
\item{...}{Parameters to be passed to the constituent NIMBLE functions
\code{\link[nimble]{configureMCMC}}, \code{\link[nimble]{runMCMC}},
\code{\link[nimble]{nimbleModel}}, \code{\link[nimble]{compileNimble}}. The
argument names will be matched to any relevant arguments in the constituent
functions. If you wish to specify that an argument is to passed to a specific
constitutent function then you can the name the argument with the form
\code{function.argument}, where 'function' is replaced with the name of the
relevant constituent function and 'argument' is replaced with the name of the
argument}

\item{warnNotFound}{A logical scalar denoting whether a warning should be
given if an argument in the \code{...} parameters has not been found in the
list of arguments in the constituent NIMBLE functions}
}
\value{
A \code{list} with named elements corresponding to each of the constituent
NIMBLE functions.  Each of these elements is a \code{list} with named elements
corresponding to each of the constituent function arguments and the unevaluated
expression to be passed to those arguments
}
\description{
A function that creates a list of parameter values that are
required for configuring, building and running the MCMC compiler for NIMBLE
and chooses sensible defaults if they are not given
}
\seealso{
\code{\link[nimble]{configureMCMC}}, \code{\link[nimble]{runMCMC}},
\code{\link[nimble]{nimbleModel}}, \code{\link[nimble]{compileNimble}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
