% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleExtensions.R
\name{nimbleScalarConstant}
\alias{nimbleScalarConstant}
\title{Create Scalar Constant for NIMBLE}
\usage{
nimbleScalarConstant(inValue, isDouble = TRUE)
}
\arguments{
\item{inValue}{A scalar value to set as a constant}

\item{isDouble}{A logical scalar.  If \code{TRUE} then the constant is a numeric scalar
and if \code{FALSE} then the constant is an integer.}
}
\value{
An object of the type returned by \code{nimbleFunction} that can be called inside
other NIMBLE code to retrieve the value of the global constant.
}
\description{
Function to create global constants that can be used in NIMBLE functions.
As of NIMBLE 0.6-12 there doesn't appear to be official support for global constants
but this instead creates functions that can be used in place of constants.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
