\name{PaGAn-package}
\alias{PaGAn-package}
\alias{PaGAn}
\docType{package}
\title{
\packageTitle{PaGAn}
}
\description{
\packageDescription{PaGAn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PaGAn}
\packageIndices{PaGAn}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PaGAn}

Maintainer: \packageMaintainer{PaGAn}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
