% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownload.R
\name{print.datasetplotGrob}
\alias{print.datasetplotGrob}
\title{Print Function for Dataset Plot Objects}
\usage{
\method{print}{datasetplotGrob}(x, ...)
}
\arguments{
\item{x}{A \code{datasetplotGrob} object to plot}

\item{...}{A set of parameter that can include: \code{newpage}, a logical
scalar that, if \code{TRUE}, plots the object on a new page}
}
\value{
A \code{datasetplotGrob} object
}
\description{
Defined method for the automatic plotting of dataset
plot objects created by the \code{\link{datasetplot}} function
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
