% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorityMesh.R
\name{priorityMesh}
\alias{priorityMesh}
\title{Creates a mesh based on gradient analysis}
\usage{
priorityMesh(
  covars = NULL,
  numCovarPoints = 100,
  ...,
  funcApply = function(curRaster, extraOptions = writeOptions) {
    
    do.call(terra::terrain, append(list(x = curRaster, v = "slope", filename =
    tempfile(fileext = ".tif")), as.list(extraOptions)))
 },
  filename = NULL,
  writeOptions = list()
)
}
\arguments{
\item{covars}{A \code{\link[terra]{SpatRaster}} object that contains the
covariate information}

\item{numCovarPoints}{An integer scalar that provides the number of extra
covariate-related triangulation points that will be produced and added to the
list of parameters appended to the \code{loc} argument before running the
\code{\link[fmesher]{fm_mesh_2d_inla}} mesh creation function.}

\item{...}{Parameters to be passed to the
\code{\link[fmesher]{fm_mesh_2d_inla}} mesh creation function}

\item{funcApply}{A function to apply to the \code{covars} object before the
priority analysis is performed.  A calue of \code{NULL} mean that no function
will be applied to the covariate information. The default option is that the
slope of the covariates will be calculated through the application of the
\code{\link[terra]{terrain}} function.}

\item{filename}{An optional character scalar giving the location to store the
calculated intensity raster object. If this value is NULL (the default) then
the raster is not stored in a file.}

\item{writeOptions}{A list of named options to be passed to the
\code{\link[terra]{writeRaster}} function when producing raster outputs.}
}
\value{
An \code{inla.mesh} object
}
\description{
A function to generate a mesh, placing points preferentially
where there are larger changes in the values of a set of environmental
covariates
}
\seealso{
\code{\link[fmesher]{fm_mesh_2d_inla}},
\code{\link[terra]{SpatRaster}}, \code{\link[terra]{terrain}},
\code{\link[terra]{writeRaster}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
