% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcInternals.R
\name{processEllipsisArgs}
\alias{processEllipsisArgs}
\title{Process ... Arguments}
\usage{
processEllipsisArgs(argsToRetrieve, ...)
}
\arguments{
\item{argsToRetrieve}{Either a function which will used to search \code{...}
for its arguments or a character vector that specifies the arguments to
retrieve}

\item{...}{Set of arguments of which some are to be retrieved in
\code{argsToRetrieve}}
}
\value{
A pairlist (as returned by \code{link[base]{alist}}) containing the
relevant arguments
}
\description{
Utility function to retrieve specified elements from a ...
argument to pass to another function
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
