% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcInternals.R
\name{processModelFormula}
\alias{processModelFormula}
\title{Process a Model Formula and Data}
\usage{
processModelFormula(inFormula, inData, centreCovs = TRUE, scaleCovs = TRUE)
}
\arguments{
\item{inFormula}{A formula object containing the model specification}

\item{inData}{A data.frame, list, or inla.data.stack object containing the
data to apply the model formula to}

\item{centreCovs}{A logical scalar denoting whether the fixed effects in the
model should be centred before the analysis: each covariate element is
subtracted by its mean. \code{centreCovs} can also be a function with one
argument that is a vector of covariate values. In this case the variable is
instead centred around the output of this function.}

\item{scaleCovs}{A logical scalar denoting whether the fixed effects in the
model should be scaled before the analysis: each covariate element is divided
by its standard deviation. \code{scaleCovs} can also be a function with one
argument that is a vector of covariate values. In this case the variable is
instead scaled around the output of this function.}
}
\value{
A list object containing the following elements:
\describe{
\item{\code{responseValues}}{The values of the response variable}
\item{\code{responseName}}{The name of the response variable}
\item{\code{hFunctions}}{Calls to special 'h' functions}
\item{\code{fFunctions}}{Calls to special 'f' functions}
\item{\code{sFunctions}}{Calls to special 's' functions}
\item{\code{covNames}}{The names of the covariates used in the model
specification and equivalent valid names to use in the NIMBLE model}
\item{\code{covFrame}}{A data frame of the processed covariates}
\item{\code{modelMatrix}}{A model matrix resulted from an expansion of the
covariates performed by \code{\link[stats]{model.matrix}}}
\item{\code{offetFrame}}{A data frame of offset terms used in the model}
}
}
\description{
Process a model formula and the input data to retrieve the
response variable and the special hierarchical components of the model
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
