% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcInternals.R
\name{processSuffix}
\alias{processSuffix}
\title{Process a Model Suffix}
\usage{
processSuffix(modelSuffix = "")
}
\arguments{
\item{modelSuffix}{A character scalar containing the model suffix to process}
}
\value{
A compliant model suffix
}
\description{
Sanity check a model suffix and convert it so that it is
compliant with BUGS
}
\seealso{
\code{\link[nimble]{nimbleCode}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
