% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleExtensions.R
\name{rbetabinStateValueMembership}
\alias{rbetabinStateValueMembership}
\title{Simulate a Variable from an Ecosystem-State Value Distribution (Beta-Binomial Error)}
\usage{
rbetabinStateValueMembership(n, stateVal, statePrec, stateProb, numTrials)
}
\arguments{
\item{n}{An integer scalar containing the number of variables to simulate (NIMBLE currently
only allows n = 1)}

\item{stateVal}{A numeric vector containing the mean values of the ecosystem state variable
associated with each alternative stable state}

\item{statePrec}{A numeric vector containing the precision (reciprocal of the variance) of the
error of the ecosystem state variable associated with each alternative stable state}

\item{stateProb}{A numeric vector containing the probabilities of being in each ecosystem stable
state (internally normalised to one)}

\item{numTrials}{A scalar containing the number of trials}
}
\value{
A scalar containing the simulated ecosystem state variable
}
\description{
Simulate a variable according to an ecosystem-state value distribution with a
beta-binomial error distribution for the ecosystem state variable.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
