% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcInternals.R
\name{responseSummary}
\alias{responseSummary}
\title{Retrieve a summary statistic of the response variable}
\usage{
responseSummary(..., summaryFunc)
}
\arguments{
\item{...}{Parameters passed to the \code{link{processModelFormula}} function
for retrieval of the relevant parameters and data.  Parameters are also
passed to the summary function.}

\item{summaryFunc}{A function to apply to the response variable}
}
\value{
A numeric scalar with the summary statistic of the response variable
}
\description{
A small utility function to retrieve a particular summary of
the response variable to help initialise set the MCMC in regression models
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
