% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSpatialINLAModel.R
\name{runINLASpatialGLM}
\alias{runINLASpatialGLM}
\title{Produce Species Distribution Models}
\usage{
runINLASpatialGLM(
  occurrenceData,
  climateData,
  specCol = NULL,
  alpha = 1.5,
  meshParameters = list(),
  meshBoundary = NULL,
  responseDensity = 100,
  outFolder = tempdir(),
  createGeoTIFF = TRUE,
  createRObFile = TRUE,
  inlaVerbose = FALSE,
  inlaKeep = FALSE,
  inlaDebug = FALSE
)
}
\arguments{
\item{occurrenceData}{A \code{\link[sf]{sf}} object containing the occurrence
data.}

\item{climateData}{A \code{\link[terra]{SpatRaster}} object containing the
climate data.}

\item{specCol}{A character scalar giving the column in the occurrence data
that represents the species.  If NULL all points are treated as one species}

\item{alpha}{A \code{numeric} scalar containing the SPDE fractional operater
order.}

\item{meshParameters}{A \code{list} containing the parameters to use in the
construction of the spatial mesh to approximate the random effects.  These
parameters are passed to the \code{\link[fmesher]{fm_mesh_2d_inla}}
and \code{\link[fmesher]{fm_nonconvex_hull_inla}} functions.}

\item{meshBoundary}{A \code{\link[sf]{sf}} object containing the boundary of
the mesh.  If this is \code{NULL} then instead construct a border polygon
using the \code{\link[fmesher]{fm_nonconvex_hull_inla}} function and
parameters passed to \code{meshParameters}.}

\item{responseDensity}{An \code{integer} scalar.  The number of sampling
locations to use when building the response curves.}

\item{outFolder}{A character scalar denoting the folder to place the SDM
output.}

\item{createGeoTIFF}{A logical scalar.  If \code{TRUE} then a series of
GeoTIFF files are created in the \code{outFolder} directory with the model
predictions for each species.}

\item{createRObFile}{A logical scalar.  If \code{TRUE} then an R object file
(with extension .rds) is created in the \code{outFolder} with the
fitted model object for each species.}

\item{inlaVerbose}{A logical scalar.  If \code{TRUE},
\code{\link[INLA]{inla}} is run in verbose mode.}

\item{inlaKeep}{A logical scalar.  If \code{TRUE}, \code{\link[INLA]{inla}}
retains working files.  These are stored in \code{outFolder}.}

\item{inlaDebug}{A logical scalar.  If \code{TRUE}, \code{\link[INLA]{inla}}
produces debugging information.}
}
\value{
A \code{list} containing two elements:
\describe{
\item{modelSummaries}{A list of \code{\link[INLA]{inla}} model objects for
each species in the \code{occurrenceData} \code{data.frame}.}
\item{spatialMesh}{A \code{SpatialPolygons}
object containing the spatial mesh used in the
\code{\link[INLA]{inla}} function.}
}
}
\description{
This function performs a generalised linear regression model with a spatial
SPDE random effect using INLA on provided occurrence data and a set of linear
covariates.
}
\author{
Joseph D. Chipperfield, \email{joseph.chipperfield@nmbu.no}
}
