% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownload.R
\name{taxonLookup}
\alias{taxonLookup}
\title{Lookup GBIF Taxonomic Information}
\usage{
taxonLookup(
  kingdom = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL,
  genus = NULL,
  species = NULL,
  other = NULL,
  limit = NULL,
  rank = NULL,
  ...
)
}
\arguments{
\item{kingdom}{A character vector containing the kingdoms that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the kingdoms specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{phylum}{A character vector containing the phyla that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the phyla specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{class}{A character vector containing the classes that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the classes specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{order}{A character vector containing the orders that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the orders specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{family}{A character vector containing the families that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the families specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{genus}{A character vector containing the genera that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the genera specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{species}{A character vector containing the species that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the species specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{other}{A character vector containing the values for the taxonomic level
specified in the \code{rank} argument that the taxa retrieved from GBIF's
taxonomies database must belong to. An entry with a length greater than one
will ensure that all retrieved entries must belong to one of the values
specified in the input. A value of \code{NULL} ensures that entries will not
be filtered according to any criteria at this taxonomic level.}

\item{limit}{An integer scalar containing the maximum number of entries to
return from the query. A value of \code{NULL} indicates that all valid
entries will be returned.}

\item{rank}{If the \code{other} parameter is non-\code{NULL} then a scalar
character must be provided here defining the taxonomic level to conduct the
query at. See \code{\link[rgbif]{name_lookup}} for acceptable values to use
for the \code{rank} argument.}

\item{...}{Other parameters passed to the \code{\link[rgbif]{name_lookup}}
function when querying GBIF's taxonomies database.}
}
\value{
A \code{data.frame} containing the taxonomic information for all
matching taxa according to the criteria specified in the function arguments.
}
\description{
A function that iteratively calls the
\code{\link[rgbif]{name_lookup}} function to retrieve information relating to
a taxon (or selection of taxa) contained at \href{https://www.gbif.org/}{GBIF}.
}
\seealso{
\code{\link[rgbif]{name_lookup}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
