% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{h}
\alias{h}
\title{Define a Hierarchical Component in a Linear Model}
\usage{
h(var, ..., model, parentSuffix = "")
}
\arguments{
\item{var}{The variable around which the hierarchical effect will be defined}

\item{...}{Named arguments to be passed to the hierarchical model
specification}

\item{model}{Either a character scalar giving the name of the
natively-supported hierarchical model to apply or a function implementing the
hierarchical model}

\item{parentSuffix}{A character scalar giving an additional suffix applied to
all elements created in the hierarchical model specification. This argument
is rarely set directly by the user but by the model definition functions
(such as \code{\link{modelDefinitionToNIMBLE}}) when generating model code}
}
\value{
A list element with the following named elements:
\describe{
\item{\code{name}}{A character scalar containing the name of the
hierarchical effect and is used as a name for intermediary variables}
\item{\code{code}}{A character scalar containing the NIMBLE code specifying
the hierarchical effect (and will be passed to
\code{\link[nimble]{nimbleCode}})}
\item{\code{constants}}{A list containing named elements corresponding to
the variables used as constants needed for the hierarchical effect in
\code{\link[nimble]{nimbleModel}}}
\item{\code{data}}{A list containing named elements corresponding to the
data nodes used for the hierarchical effect in
\code{\link[nimble]{nimbleModel}}}
\item{\code{inits}}{A named list of starting values for model variables used
in the hierarchical effect and passed to \code{\link[nimble]{nimbleModel}}}
\code{monitors}{The nodes of the hierarchical effect to
 monitor in the MCMC and passed to \code{\link[nimble]{configureMCMC}}}
\code{monitors2}{The nodes of the hierarchical effect to
 monitor in the supplemental chain monitor in the MCMC and passed to
 \code{\link[nimble]{configureMCMC}}}
}
}
\description{
Function to specify a hierarchical component with a Bayesian
model specification
}
\seealso{
\code{\link[nimble]{nimbleCode}},
\code{\link[nimble]{configureMCMC}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
