% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcInternals.R
\name{modelDefinitionToNIMBLE}
\alias{modelDefinitionToNIMBLE}
\title{Convert a Model Definition into NIMBLE Code}
\usage{
modelDefinitionToNIMBLE(
  formula,
  data,
  family = "gaussian",
  link = "identity",
  centreCovs = TRUE,
  scaleCovs = TRUE,
  suffix = "",
  sharedTermNode = NULL,
  ...
)
}
\arguments{
\item{formula}{A formula that takes a similar form to that used in
\code{\link[stats]{glm}} except that hierarchical components can be added
to the model specification using the \code{\link{h}} function.  In addition
it is possible to specify random effects from INLA using the
\code{\link[INLA]{f}} function.}

\item{data}{A data.frame, list, or inla.data.stack object containing the
data to apply the model formula to.}

\item{family}{The specification of the likelihood family to use. The default
is a Gaussian distribution with identity link. The argument can either be
a string containing the name of the likelihood family (case insensitive) or
a \code{\link[stats]{family}} object containing the family and link function
specification. If the \code{family} argument is a string then the \code{link}
argument can be used to specify an alternative link function than the default
link function for that distribution. See \code{names(errorFamilites())} for a
list of natively supported error distributions. Alternatively, the user can
specify a custom error distribution through the application of the
\code{\link{customError}} function.}

\item{link}{A character scalar giving the name of the link function to be
used with the likelihood family. See \code{unique(unlist(errorFamilies()))}
for a list of natively-supported link functions.}

\item{centreCovs}{A logical scalar denoting whether the fixed effects in the
model should be centred before the analysis: each covariate element is
subtracted by its mean. \code{centreCovs} can also be a function with one
argument that is a vector of covariate values. In this case the variable is
instead centred around the output of this function.}

\item{scaleCovs}{A logical scalar denoting whether the fixed effects in the
model should be scaled before the analysis: each covariate element is divided
by its standard deviation. \code{scaleCovs} can also be a function with one
argument that is a vector of covariate values. In this case the variable is
instead scaled around the output of this function.}

\item{suffix}{A character scalar that will be appended to all variables used
in the NIMBLE code (including constants and data)}

\item{sharedTermNode}{A character containing the name of a node containing
the expectation that is shared across different likelihoods. Defaults to
\code{NULL} and this default should be used in most cases where the
model specification isn't being combined with a larger model component}

\item{...}{Other parameters that are passed to the error distribution
processing functions}
}
\value{
A list object containing the following named elements:
\describe{
\item{\code{responseValues}}{The values of the response variable}
\item{\code{responseName}}{The name of the response variable}
\item{\code{hFunctions}}{Calls to special 'h' functions}
\item{\code{fFunctions}}{Calls to special 'f' functions}
\item{\code{sFunctions}}{Calls to special 's' functions}
\item{\code{covNames}}{The names of the covariates used in the model
specification and equivalent valid names to use in the NIMBLE model}
\item{\code{covFrame}}{A data frame of the processed covariates}
\item{\code{modelMatrix}}{A model matrix resulted from an expansion of the
covariates performed by \code{\link[stats]{model.matrix}}}
\item{\code{offetFrame}}{A data frame of offset terms used in the model}
\item{\code{code}}{NIMBLE model specification code for the model
defined using the input parameters (see \code{\link[nimble]{nimbleCode}})}
\item{\code{constants}}{A list of constants to be passed to NIMBLE (see
\code{\link[nimble]{nimbleModel}})}
\item{\code{data}}{A list of data to be passed to NIMBLE (see
\code{\link[nimble]{nimbleModel}})}
\item{\code{inits}}{A list of starting values for model variables to be
passed to NIMBLE (see \code{\link[nimble]{nimbleModel}})}
\item{\code{dimensions}}{Named list of dimensions for variables to be
passed to NIMBLE (see \code{\link[nimble]{nimbleModel}})}
\item{\code{monitors}}{A character vector of names of variables to record
during MCMC sampling in NIMBLE (see \code{\link[nimble]{configureMCMC}})}
\item{\code{monitors2}}{A character vector of names of varaibles to record
during MCMC sampling in NIMBLE with a different thinning interval (see
\code{\link[nimble]{configureMCMC}})}
\item{\code{hierAttributes}}{A list containing attributes that are set
in each of the hierarchical component models}
\item{\code{link}}{A link list structure (as returned from
\code{\link{customLink}}) that defines the link function used in the model}
\item{\code{family}}{A error distribution list structure (as returned from
\code{\link{customError}}) that defines the error distribution used in the
model}
}
}
\description{
A function to create a full NIMBLE model definition from a
standard model specification with a syntax similar to that used in the
\code{\link[stats]{glm}} model specification
}
\seealso{
\code{\link[nimble]{nimbleCode}} \code{\link{customLink}}
\code{\link{customError}} \code{\link{errorFamilies}}
\code{\link[stats]{glm}} \code{\link[INLA]{inla.stack.data}}
\code{\link[stats]{model.matrix}} \code{\link[nimble]{nimbleModel}}
\code{\link[nimble]{configureMCMC}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
